+++ app/Models/Member.php
@@ -11,6 +11,7 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use Illuminate\Notifications\Notifiable;
 use Illuminate\Support\Arr;
 
 /**
@@ -39,6 +40,7 @@ class Member extends Model
 {
     use HasFactory;
     use SoftDeletes;
+    use Notifiable;
 
     public const MEMBER_EDIT_PERMISSION = 'memberEdit';